// src/app/modules/common-regobs-api/model/danger-obs-edit-model.ts
/**
 * RegObs API, build: Local build - commit: ec80925f
 *
 * Contact: regobs@nve.no
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DangerObsEditModel { 
    /**
     * Sett naturfare. Tabellen brukes av alle naturfarer (snø, jord, vann, is). The GeoHazardKD unique identifier
     */
    GeoHazardTID?: number | null;
    /**
     * Faretegn er listet i tabellen DangerSignKD. The DangerSignKD unique identifier
     */
    DangerSignTID?: number | null;
    /**
     * Kommentarfelt for å skrive utfyllende tekst om faretegnet.
     */
    Comment?: string | null;
}