// src/app/modules/common-regobs-api/model/snow-surface-edit-model.ts
/**
 * RegObs API, build: 20251103.5 - commit: 62e967e0
 *
 * Contact: regobs@nve.no
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SnowSurfaceEditModel { 
    /**
     * Total snødybde i cm. Vi ønsker et gjennomsnitt for området.
     */
    SnowDepth?: number | null;
    /**
     * Nysnødybde i cm. Vi ønsker et gjennomsnitt for området
     */
    NewSnowDepth24?: number | null;
    /**
     * Nysnøgrensa. Hvor faller nedbør som snø? Meter over havet.
     */
    NewSnowLine?: number | null;
    /**
     * Overflatefuktighet
     */
    SurfaceWaterContentTID?: number | null;
    /**
     * Hvor mye har vinden transportert snøen? Valg fra nedtrekksmeny. The SnowDriftKD unique identifier
     */
    SnowDriftTID?: number | null;
    /**
     * Snødekkehardhet
     */
    SnowSurfaceTID?: number | null;
    /**
     * Skiføre
     */
    SkiConditionsTID?: number | null;
    /**
     * Kommentarfelt for å skrive utfyllende tekst om observasjonen.
     */
    Comment?: string | null;
    /**
     * Grense lagdelt snø (moh)
     */
    HeightLimitLayeredSnow?: number | null;
    /**
     * Snøgrense (moh)
     */
    SnowLine?: number | null;
}